+++ src/lib.rs
@@ -85,6 +85,7 @@ pub mod window_size;
 // HTTP PROTOCOL MODULES (depends on TCP)
 // ============================================================================
 pub mod http;
+pub mod http1_process;
 pub mod http_languages;
 pub mod http_process;
 mod observable_http_signals_matching;
